Usa ints ao inves de BigIntegers - #443

+++ src/main/java/br/gov/servicos/v3/schema/Ate.java
@@ -8,7 +8,6 @@
 
 package br.gov.servicos.v3.schema;
 
-import java.math.BigInteger;
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
 import javax.xml.bind.annotation.XmlAttribute;
@@ -38,31 +37,23 @@ import javax.xml.bind.annotation.XmlType;
 public class Ate {
 
     @XmlAttribute(name = "max", required = true)
-    protected BigInteger max;
+    protected int max;
     @XmlAttribute(name = "unidade", required = true)
     protected TipoDeTempoTotalEstimado unidade;
 
     /**
      * Gets the value of the max property.
      * 
-     * @return
-     *     possible object is
-     *     {@link BigInteger }
-     *     
      */
-    public BigInteger getMax() {
+    public int getMax() {
         return max;
     }
 
     /**
      * Sets the value of the max property.
      * 
-     * @param value
-     *     allowed object is
-     *     {@link BigInteger }
-     *     
      */
-    public void setMax(BigInteger value) {
+    public void setMax(int value) {
         this.max = value;
     }
